Batman can now attack his enemies.
When he attacks the Penguin, we get a modal reaction.

Here's my thought process on adding this feature:

- There needs to be a button on the Batman view, that the user can
  press, that means "Attack current villain."
- The Batman component doesn't know about villains, so it can't be an
  action of that component.
- So it will have to be a parameter. Batman's parent* will have to pass
  in the action for us to use.
- Should it be a Message or an Attribute? Message. Attribute is too vague.
- Code that much in. Follow the compiler errors.

Done.

FWIW, I think of button clicks and such as as, "the user wants to convey
the intention that ...". Once you realise that the `Attack` intention
has to come from higher up in the tree, so it must be a parameter, the
rest sort of falls out. You just keep going up the model tree until you can
actually handle that intention, make it an `Action` of that level, and
then complete all the wiring.

* Batman has no parents.